toggle for networks and padding fixed
