+++ src/components/Header/header.css
@@ -5,8 +5,8 @@
 
 .menuBody {
   background-color: #48494F !important;
-  padding-top: 10px !important;
-  padding-bottom: 10px !important;
+  padding-top: 5px !important;
+  padding-bottom: 5px !important;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
@@ -14,8 +14,8 @@
 
 .listItem {
   color: #fff !important;
-  padding-left: 46px !important;
-  padding-right: 46px !important;
+  padding-left: 44px !important;
+  padding-right: 44px !important;
   padding-top: 10px !important;
   padding-bottom: 10px !important;
   font-size: 14px !important;
@@ -30,7 +30,7 @@
 }
 
 .listItem:last-child {
-  margin-bottom: 0px;
+  margin-bottom: 0px !important;
 }
 
 .listItem:hover {
